mimxrt uart: Get rid of machine_uart.h.
