--- ports/mimxrt/machine_uart.h
@@ -1,49 +1,0 @@
-/*
- * This file is part of the MicroPython project, http://micropython.org/
- *
- * The MIT License (MIT)
- *
- * Copyright (c) 2021 Damien P. George
- * Copyright (c) 2021 Robert Hammelrath
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-#ifndef MICROPY_INCLUDED_MIMXRT_UART_H
-#define MICROPY_INCLUDED_MIMXRT_UART_H
-
-#include "fsl_common.h"
-#include "fsl_lpuart.h"
-#include "fsl_iomuxc.h"
-
-typedef struct _machine_uart_obj_t {
-    mp_obj_base_t base;
-    struct _lpuart_handle handle;
-    lpuart_config_t config;
-    LPUART_Type *lpuart;
-    uint16_t timeout;       // timeout waiting for first char (in ms)
-    uint16_t timeout_char;  // timeout waiting between chars (in ms)
-    uint8_t id;
-    uint8_t invert;
-    bool new;
-} machine_uart_obj_t;
-
-extern const mp_obj_type_t machine_uart_type;
-extern bool lpuart_set_iomux(int8_t uart); 
-
-#endif // MICROPY_INCLUDED_MIMXRT_UART_H
