+++ ports/mimxrt/modmachine.h
@@ -32,6 +32,7 @@
 extern const mp_obj_type_t machine_adc_type;
 extern const mp_obj_type_t machine_timer_type;
 extern const mp_obj_type_t machine_rtc_type;
+extern const mp_obj_type_t machine_uart_type;
 
 void machine_adc_init(void);
 void machine_timer_init_PIT(void);
